change_request/requirements.txt
alembic==1.6.5
blinker==1.4
click==8.0.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==2.0.1
Flask-Bootstrap4==4.0.2
Flask-DebugToolbar==0.11.0
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
importlib-metadata==4.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.17
typing-extensions==3.10.0.0
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.4.1
